+++ examples/scripts/binary_classifier.py
@@ -19,28 +19,28 @@ y = df["income"]
 X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
 
 automl = AutoML(
-    #results_path="AutoML_30",
-    algorithms=["Random Forest"],
+    results_path="AutoML_38",
+    #algorithms=["Random Forest"],
     total_time_limit=20,
-    explain_level=0,
+    #explain_level=0,
     # validation={"validation_type": "split"},
-    mode="Explain",
+    #mode="Explain",
     # validation={"validation_type": "split"}
-    validation={
-        "validation_type": "kfold",
-        "k_folds": 2,
-        "shuffle": True,
-        "stratify": True,
-    },
-    golden_features=True,
-    feature_selection=True
+    #validation_starategy={
+    #    "validation_type": "kfold",
+    #    "k_folds": 2,
+    #    "shuffle": True,
+    #    "stratify": True,
+    #},
+    #golden_features=True,
+    #feature_selection=True
 )
-automl.set_advanced(
-    start_random_models=20, hill_climbing_steps=10, top_models_to_improve=3
-)
+#automl.set_advanced(
+# #   start_random_models=20, hill_climbing_steps=10, top_models_to_improve=3
+#)
 automl.fit(X_train, y_train)
 
-predictions = automl.predict(X_test)
+predictions = automl.predict_all(X_test)
 
 print(predictions.head())
 print(predictions.tail())
