+++ examples/scripts/multi_class_classifier.py
@@ -19,7 +19,7 @@ y = df["class"]
 
 automl = AutoML(
     # results_path="AutoML_41",
-    algorithms=["Random Forest"],
+    #algorithms=["Random Forest"],
     # algorithms=["Neural Network"],
     #    "Linear",
     #    "Xgboost",
@@ -28,7 +28,7 @@ automl = AutoML(
     # total_time_limit=100,
     # tuning_mode="Normal",
     # explain_level=0,
-    mode="Perform"
+    #mode="Perform"
 )
 # automl.set_advanced(start_random_models=1)
 automl.fit(X, y)
